requirements.txt
bcrypt==4.2.0
black==24.10.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
flake8==7.1.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
paramiko==3.5.0
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyNaCl==1.5.0
python-dotenv==1.0.1
